添加统计磁盘信息功能

+++ src/main/java/com/xy/models/ShopWallet.java
@@ -1,6 +1,7 @@
 package com.xy.models;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Table(name = "shop_wallet")
 public class ShopWallet {
@@ -28,12 +29,12 @@ public class ShopWallet {
     /**
      * 余额
      */
-    private Integer money;
+    private BigDecimal money;
 
     /**
      * 积分余额
      */
-    private Integer score;
+    private BigDecimal score;
 
     /**
      * 支付使用的银行卡账户
@@ -126,7 +127,7 @@ public class ShopWallet {
      *
      * @return money - 余额
      */
-    public Integer getMoney() {
+    public BigDecimal getMoney() {
         return money;
     }
 
@@ -135,7 +136,7 @@ public class ShopWallet {
      *
      * @param money 余额
      */
-    public void setMoney(Integer money) {
+    public void setMoney(BigDecimal money) {
         this.money = money;
     }
 
@@ -144,7 +145,7 @@ public class ShopWallet {
      *
      * @return score - 积分余额
      */
-    public Integer getScore() {
+    public BigDecimal getScore() {
         return score;
     }
 
@@ -153,7 +154,7 @@ public class ShopWallet {
      *
      * @param score 积分余额
      */
-    public void setScore(Integer score) {
+    public void setScore(BigDecimal score) {
         this.score = score;
     }
